Rename component to metrics_collection_traffic
So it would be more clear that this component handles metrics
collection traffic, and does not collect metrics
