+++ src/operator/controllers/metrics_collection_traffic/pod_reconciler.go
@@ -1,45 +1,50 @@
-package metrics_collectors
+package metrics_collection_traffic
 
 import (
 	"context"
 	"github.com/otterize/intents-operator/src/shared/errors"
 	"github.com/otterize/intents-operator/src/shared/injectablerecorder"
 	"github.com/samber/lo"
-	v1 "k8s.io/api/networking/v1"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/client-go/tools/record"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 )
 
+//+kubebuilder:rbac:groups="",resources=pods,verbs=get;list;watch
 //+kubebuilder:rbac:groups="networking.k8s.io",resources=networkpolicies,verbs=get;update;patch;list;watch;delete;create
 
-type NetworkPolicyReconciler struct {
+type PodReconciler struct {
 	client.Client
 	netpolHandle *NetworkPolicyHandler
 	injectablerecorder.InjectableRecorder
 }
 
-func NewNetworkPolicyReconciler(
-	client client.Client,
-	netpolHandle *NetworkPolicyHandler) *NetworkPolicyReconciler {
-	return &NetworkPolicyReconciler{
+func NewPodReconciler(client client.Client, netpolHandle *NetworkPolicyHandler) *PodReconciler {
+	return &PodReconciler{
 		Client:       client,
 		netpolHandle: netpolHandle,
 	}
 }
 
-func (r *NetworkPolicyReconciler) SetupWithManager(mgr ctrl.Manager) error {
+func (r *PodReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	recorder := mgr.GetEventRecorderFor("intents-operator")
 	r.InjectRecorder(recorder)
 
 	return ctrl.NewControllerManagedBy(mgr).
-		For(&v1.NetworkPolicy{}).
+		For(&corev1.Pod{}).
 		WithOptions(controller.Options{RecoverPanic: lo.ToPtr(true)}).
 		Complete(r)
 }
 
-func (r *NetworkPolicyReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
-	err := r.netpolHandle.HandleAllServicesInNamespace(ctx, req.Namespace)
+func (r *PodReconciler) InjectRecorder(recorder record.EventRecorder) {
+	r.Recorder = recorder
+	r.netpolHandle.InjectRecorder(recorder)
+}
+
+func (r *PodReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
+	err := r.netpolHandle.HandleAllPodsInNamespace(ctx, req.Namespace)
 	if err != nil {
 		return ctrl.Result{}, errors.Wrap(err)
 	}
