+++ src/operator/controllers/metrics_collection_traffic/service_reconciler.go
@@ -1,4 +1,4 @@
-package metrics_collectors
+package metrics_collection_traffic
 
 import (
 	"context"
